First implementation of rex debug viewer (#11)
The rex-viewer is intended to be a debug viewer and does not reflect the same result as with our AR-based viewers. This viewer is only meant for debugging.

There are still some features missing, and currently only Linux is tested as platform. For instance, material coloring is not yet supported.